requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
python-decouple==3.8
mysqlclient==2.2.0
django-cors-headers==4.3.1
